refactor: extract codex hook adapter helpers
