+++ .safeword/logs/explore-codex-open-tickets.md
@@ -1,0 +1,7 @@
+# Work Log: explore Codex open tickets
+
+## 2026-06-24
+
+- 2026-06-24T02:00:24Z Started: User asked to catch up to main, inspect open Codex-related tickets, and tackle Codex bugs where practical.
+- 2026-06-24T02:00:24Z Found: Worktree is clean on `main`; `git pull --ff-only origin main` reported already up to date.
+- 2026-06-24T02:03:24Z Selected: `W0E292` as the actionable open Codex task. Created branch `codex/w0e292-hook-adapter-helpers`. Stale-index findings: `HMZSCD`, `K2ZP40`, and `Y5GS4X` are marked done in their ticket files despite appearing in the index as in progress.
